+++ .eslintrc.js
@@ -2,23 +2,28 @@ module.exports = {
   root: true,
   parser: 'babel-eslint',
   parserOptions: {
-    sourceType: 'module'
+    sourceType: 'module',
+    ecmaFeatures: {
+      'jsx': true
+    },
   },
-  // https://github.com/feross/standard/blob/master/RULES.md#javascript-standard-style
-  extends: 'standard',
-  // required to lint *.vue files
   plugins: [
-    'html'
+    'react'
   ],
-  // add your custom rules here
-  'rules': {
+  extends: 'react-app',
+  rules: {
+    'semi': [2, 'never'],
     // allow paren-less arrow functions
     'arrow-parens': 0,
     // allow async-await
-    'generator-star-spacing': 0,
+    'generator-star-spacing': 2,
     // allow debugger during development
-    'no-debugger': process.env.NODE_ENV === 'production' ? 2 : 0,
+    'no-debugger': 2,
     'comma-dangle': 0,
     'camelcase': 0,
+    'no-alert': 2,
+    'space-before-function-paren': 2,
+    'react/jsx-uses-react': 2,
+    'react/jsx-uses-vars': 2,
   }
 }
