+++ old-vue/src/components/con-details.vue
