+++ old-vue/test/unit/specs/con-details.js
