+++ package.json
@@ -14,78 +14,26 @@
   },
   "license": "MIT",
   "scripts": {
-    "start": "node build/dev-server.js",
-    "build": "node build/build.js",
-    "test": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js --single-run",
-    "testlive": "cross-env BABEL_ENV=test karma start test/unit/karma.conf.js",
-    "lint": "eslint --ext .js,.vue src test/unit/specs"
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test --env=jsdom",
+    "eject": "react-scripts eject",
+    "lint": "eslint src",
+    "lintfix": "eslint src --fix"
   },
   "engines": {
-    "node": ">= 6.0.0",
-    "npm": ">= 3.0.0"
+    "node": "8.x",
+    "npm": "5.x",
+    "yarn": "1.3.x"
   },
   "dependencies": {
-    "autoprefixer": "^6.7.6",
-    "babel-core": "^6.23.1",
-    "babel-eslint": "^7.0.0",
-    "babel-loader": "^6.4.0",
-    "babel-plugin-istanbul": "^4.0.0",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-stage-2": "^6.22.0",
-    "babel-register": "^6.23.0",
-    "chai": "^3.5.0",
-    "chalk": "^1.1.3",
-    "cross-env": "^3.2.3",
-    "css-loader": "^0.26.4",
-    "eslint": "^3.17.1",
-    "eslint-config-standard": "^7.0.1",
-    "eslint-friendly-formatter": "^2.0.5",
-    "eslint-loader": "^1.6.3",
-    "eslint-plugin-html": "^2.0.1",
-    "eslint-plugin-promise": "^3.5.0",
-    "eslint-plugin-standard": "^2.1.1",
-    "eventsource-polyfill": "^0.9.6",
-    "expose-loader": "^0.7.3",
-    "express": "^4.15.2",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "friendly-errors-webpack-plugin": "^1.6.1",
-    "function-bind": "^1.0.2",
-    "html-webpack-plugin": "^2.28.0",
-    "inject-loader": "^2.0.1",
-    "json-loader": "^0.5.4",
-    "karma": "^1.5.0",
-    "karma-coverage": "^1.1.1",
-    "karma-mocha": "^1.2.0",
-    "karma-phantomjs-launcher": "^1.0.0",
-    "karma-sinon-chai": "^1.2.4",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-spec-reporter": "0.0.30",
-    "karma-webpack": "^1.7.0",
-    "lodash": "^4.17.4",
-    "lolex": "^1.6.0",
-    "marked": "^0.3.6",
-    "mocha": "^3.1.0",
-    "node-sass": "^4.5.1",
-    "opn": "^4.0.2",
-    "phantomjs-prebuilt": "^2.1.3",
-    "raven-js": "^3.12.1",
-    "sass-loader": "^4.1.1",
-    "semver": "^5.3.0",
-    "shelljs": "^0.7.6",
-    "sinon": "^1.17.3",
-    "sinon-chai": "^2.8.0",
-    "style-loader": "^0.16.0",
-    "url-loader": "^0.5.8",
-    "vue": "^2.2.1",
-    "vue-loader": "^11.1.4",
-    "vue-multiselect": "^2.0.0-beta.15",
-    "vue-router": "^2.3.0",
-    "vue-template-compiler": "^2.2.1",
-    "webpack": "^1.13.2",
-    "webpack-dev-middleware": "^1.10.1",
-    "webpack-hot-middleware": "^2.17.1",
-    "webpack-merge": "^0.14.1"
+    "marked": "^0.3.7",
+    "node-sass": "^4.7.2",
+    "raven-js": "^3.20.1",
+    "react": "^16.2.0",
+    "react-app-rewired": "^1.3.8",
+    "react-dom": "^16.2.0",
+    "react-scripts": "1.0.17",
+    "sass-loader": "^6.0.6"
   }
 }
