+++ public/index.html
@@ -2,7 +2,7 @@
 <html>
   <head>
     <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1, maximum-scale=1, user-scalable=0"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     <title>tutorcruncher-socket</title>
     <style>
       @import url(https://fonts.googleapis.com/css?family=Open+Sans:400,500);
