+++ src/styles/subject-select.scss
@@ -1,4 +1,4 @@
-@import './vue-multiselect/dist/vue-multiselect.min';
+//@import './vue-multiselect/dist/vue-multiselect.min';
 @import '../conf';
 
 .multiselect__select {
